+++ package-testing/sdk-test-runner/src/protocol/FeatureNotSupportedError.ts
@@ -1,0 +1,6 @@
+export class FeatureNotSupportedError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = 'FeatureNotSupportedError';
+  }
+}
