+++ package-testing/sdk-test-runner/src/protocol/SDKConnectionFailure.ts
@@ -1,0 +1,3 @@
+export type SDKConnectionFailure = {
+  errorMessage?: string;
+};
