+++ package-testing/sdk-test-runner/src/protocol/SDKInfo.ts
@@ -1,0 +1,4 @@
+export type SDKInfo = {
+  sdkName: string;
+  supportsBandits: boolean;
+};
