+++ package-testing/sdk-test-runner/src/protocol/constants.ts
@@ -1,0 +1,3 @@
+export const ASSIGNMENT_PATH = '/flags/v1/assignment';
+export const BANDIT_PATH = '/bandits/v1/action';
+export const RESET_PATH = '/sdk/reset';
