+++ package-testing/sdk-test-runner/src/protocol/index.ts
@@ -1,0 +1,8 @@
+import { ClientSDKRelay } from './ClientSDKRelay';
+import { FeatureNotSupportedError } from './FeatureNotSupportedError';
+import { SDKConnectionFailure } from './SDKConnectionFailure';
+import { SDKInfo } from './SDKInfo';
+import { SDKRelay } from './SDKRelay';
+import { ServerSDKRelay } from './ServerSDKRelay';
+
+export { ClientSDKRelay, ServerSDKRelay, SDKInfo, SDKRelay, SDKConnectionFailure, FeatureNotSupportedError };
